impl struct

+++ fascript/fascript.hpp
@@ -14,9 +14,9 @@ public:
 
 private:
 	std::mutex m_mtx;
-	std::map<std::string, size_t> m_name_to_id;
-	std::map<size_t, Value> m_id_to_var;
-	size_t m_next_id;
+	std::map<std::string, uint16_t> m_name_to_id;
+	std::map<uint16_t, Value> m_id_to_var;
+	uint16_t m_next_id;
 };
 }
